+++ packages/web-js-sdk/src/index.umd.ts
@@ -1,8 +1,4 @@
-import {
-  REFRESH_TOKEN_KEY,
-  SESSION_TOKEN_KEY,
-} from './enhancers/withPersistTokens/constants';
-import sdk from './decoratedSdk';
+import sdk, { REFRESH_TOKEN_KEY, SESSION_TOKEN_KEY } from './index';
 
 // We export the tokens constant in this way so that umd bundles will have the constants as properties of the default export
 // But still can use the default export as a function (e.g. Descope({ projectId: 'pid'}))
