+++ packages/web-js-sdk/test/umd.test.ts
@@ -1,14 +1,18 @@
-import createSdk from '../src/index.umd';
+// We run build before test, so we can import from dist
+import Descope from '../dist/index.umd.js';
+
+const mockFetch = jest.fn().mockReturnValueOnce(new Promise(() => {}));
+global.fetch = mockFetch;
 
 describe('umd sdk', () => {
   it('should export sdk', async () => {
-    expect(createSdk).toBeDefined();
-    expect(createSdk).toBeInstanceOf(Function);
-    expect(createSdk({ projectId: 'pid' })).not.toThrow();
+    expect(Descope).toBeDefined();
+    expect(Descope).toBeInstanceOf(Function);
+    expect(() => Descope({ projectId: 'pid' })).not.toThrow();
   });
 
   it('should export constants', () => {
-    expect(createSdk['SESSION_TOKEN_KEY']).toBeDefined();
-    expect(createSdk['REFRESH_TOKEN_KEY']).toBeDefined();
+    expect(Descope['SESSION_TOKEN_KEY']).toBeDefined();
+    expect(Descope['REFRESH_TOKEN_KEY']).toBeDefined();
   });
 });
